+++ feincms3/__init__.py
@@ -1,4 +1,4 @@
 from __future__ import absolute_import, unicode_literals
 
-VERSION = (0, 3, 0)
+VERSION = (0, 4, 0)
 __version__ = '.'.join(map(str, VERSION))